More refactoring and shit
